More multiple package source goodness
pacman.py now done. Additionally, common functions needed for
pkg.install have been broken out into pkg_resource.py. zypper.py has
been updated to use the functions in pkg_resource.py.

Yay, code re-use!
